zeroize: Rewrite using 'core::ptr::write_volatile'
See the newly added documentation in `lib.rs` for the rationale.

This commit deletes all of the platform-specific boilerplate for
leveraging various OS-specific APIs and replaces it all with
`write_volatile`.

This not only provides a massive simplification, but a solid foundation
for supporting zeroization of all of the core Rust types.
