--- zeroize/src/os/windows/explicit_bzero_shim.c
@@ -1,11 +1,0 @@
-/**
- * `explicit_bzero()` shim for invoking `SecureZeroMemory` on Windows, which
- * is provided as a macro.
- */
-
-#include <windows.h>
-#include <wincrypt.h>
-
-void explicit_bzero(void *dest, size_t n) {
-    SecureZeroMemory(dest, n);
-}
